Refactor analytics constants using part files - 1
