+++ packages/devtools_app/lib/src/shared/analytics/constants/_cpu_profiler_constants.dart
@@ -1,0 +1,18 @@
+// Copyright 2023 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+part of '../constants.dart';
+
+enum CpuProfilerEvents {
+  profileGranularity,
+  loadAllCpuSamples,
+  profileAppStartUp,
+  cpuProfileFlameChartHelp,
+  cpuProfileProcessingTime,
+  cpuProfileDisplayTreeGuidelines,
+}
+
+enum CpuProfilerDocs {
+  profileGranularityDocs,
+}
